refactor(nodes): extract class-name management from Element into ElementClasses
